All animations of the player have been corrected

+++ The Cliffborn Islands/Motor2D/j1Player.h
@@ -64,7 +64,7 @@ public:
 	int horizontalSpeed;
 	int verticalSpeed;
 	int initialVerticalSpeed;
-	float gravity;
+	int gravity;
 
 	uint currentTime;
 	uint lastTime;
